deleted myApp after copying them into my branches version of AlechemEcommerce
