+++ FrontEnd/AlchemEcommerce/src/app/app-routing.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { CheckoutComponent } from './components/checkout/checkout.component';
 import { LoginComponent } from './components/login/login.component';
 import { RegisterComponent } from './components/register/register.component';
 
@@ -12,6 +13,11 @@ const routes: Routes = [
     path: 'register',
     component: RegisterComponent,
   },
+
+  {
+    path: 'checkout',
+    component: CheckoutComponent,
+  }
 ];
 
 @NgModule({
