--- FrontEnd/myApp/src/app/app-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { LoginComponentComponent } from './components/login-component/login-component.component';
-
-const routes: Routes = [
-  {
-    path: 'login',
-    component: LoginComponentComponent,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule],
-})
-export class AppRoutingModule {}
