--- FrontEnd/myApp/src/app/app.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { LoginComponentComponent } from './components/login-component/login-component.component';
-
-@NgModule({
-  declarations: [AppComponent, LoginComponentComponent],
-  imports: [BrowserModule, AppRoutingModule],
-  providers: [],
-  bootstrap: [AppComponent],
-})
-export class AppModule {}
