--- FrontEnd/myApp/src/app/components/login-component/login-component.component.html
@@ -1,22 +1,0 @@
-<body>
-
-
-    <div class="container"> <video onloadedmetadata="this.muted=true" autoplay loop id="myVid">
-            <source src="../../../assets/videos/Fire_41___45s___4k_res.mp4" type="video/mp4">
-        </video> </div>
-    <div class="div" id="loginDiv">
-        <h1>Login</h1>
-
-        <label for="username">Username</label>
-        <input type="text" name="user" class="input-box" id="username-input">
-
-        <br><br>
-
-        <label for="password">Password</label>
-        <input type="password" name="pass" class="input-box" id="password-input">
-
-        <button class="btn" type="button" id="loginBtn">Login</button>
-
-    </div>
-
-</body>
