--- FrontEnd/myApp/src/app/components/login-component/login-component.component.ts
@@ -1,14 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-login-component',
-  templateUrl: './login-component.component.html',
-  styleUrls: ['./login-component.component.css'],
-})
-export class LoginComponentComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
-
-function login() {}
