--- FrontEnd/myApp/src/styles.css
@@ -1,15 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
-@font-face {
-  font-family: "dum2";
-  src: local(dum2) url(./assets/Fonts/dum2/dum2.ttf);
-}
-div {
-  border-radius: 10px;
-  padding: 15px;
-  background-color: goldenrod;
-  opacity: 0.6;
-  font-family: "dum2";
-}
-.hide {
-  display: none;
-}
